+++ .eslintrc.json
@@ -3,5 +3,8 @@
     "env": {
         "node": true,
         "jest": true
+    },
+    "rules":{
+        "linebreak-style": 0
     }
 }
