+++ package.json
@@ -14,5 +14,8 @@
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.8.0",
     "jest": "^22.4.3"
+  },
+  "rules": {
+    "linebreak-style": 0
   }
 }
