--- src/components/DivElement.js
@@ -1,13 +1,0 @@
-const HTMLElement = require('./HTMLElement');
-
-// Define class here
-class DivElement extends HTMLElement{
-    constructor(content){
-        super(content);
-        this.tag = "div";
-        this.content = content;
-    }
-}
-
-// Export class here
-module.exports = DivElement;
