+++ src/components/HTMLElement 2.js
@@ -1,0 +1,14 @@
+// Define class here
+class HTMLElement {
+  constructor(tag, content) {
+    this.tag = tag;
+    this.content = content;
+  }
+
+  render() {
+    return `'<' + ${this.tag} + '>' + ${this.content} + '</' + ${this.tag} + '>'`;
+  }
+}
+
+// Export class here
+module.exports = HTMLElement;
