--- src/components/HTMLElement.js
@@ -1,14 +1,0 @@
-// Define class here
-class HTMLElement {
-    constructor(tag, content){
-        this.tag = tag;
-        this.content = content;
-    }
-
-    render(){
-        return "<" + this.tag + ">" + this.content + "</" + this.tag + ">";
-    }
-}
-
-// Export class here
-module.exports = HTMLElement;
